component layout fixes
