+++ apps/admin/components/logistics/onboarding-review/index.ts
@@ -1,0 +1,4 @@
+export * from "./OnboardingDetail";
+export * from "./OnboardingHeader";
+export * from "./OnboardingList";
+export * from "./OnboardingLoading";
