+++ apps/admin/components/logistics/onboarding/index.ts
@@ -1,0 +1,2 @@
+export * from "./ExecProfile";
+export * from "./General";
