--- apps/admin/components/onboarding/index.ts
@@ -1,6 +1,0 @@
-export * from "./ExecProfile";
-export * from "./General";
-export * from "./hiring/OnboardingDetail";
-export * from "./hiring/OnboardingHeader";
-export * from "./hiring/OnboardingList";
-export * from "./hiring/OnboardingLoading";
